pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["pms <apark0907@ajou.ac.kr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "~4.0"
black = "^22.12.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
dj-database-url = "^1.3.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
